+++ pykit/err_utils.py
@@ -6,7 +6,7 @@ from pykit.obj import get_fqname
 from pykit.res import Err, Ok, Res
 
 
-def try_get_err_traceback_str(err: Exception) -> str | None:
+def get_err_traceback_str(err: Exception) -> str | None:
     s = None
     tb = err.__traceback__
     if tb:
@@ -20,7 +20,7 @@ def try_get_err_traceback_str(err: Exception) -> str | None:
 def create_err_dto(err: Exception) -> Res[ErrDto]:
     name = get_fqname(err)
     msg = ", ".join([str(a) for a in err.args])
-    stacktrace = try_get_err_traceback_str(err)
+    stacktrace = get_err_traceback_str(err)
     errcode_res = Code.get_from_type(type(err))
     if isinstance(errcode_res, Err):
         return errcode_res