+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - added Code.get_regd_codeid_by_type()
 - removed io.py
 - added log.track and log.atrack
+- added res.track and res.atrack
 
 # 0.9.2